% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\docType{data}
\name{.syn}
\alias{.syn}
\title{Main synapseclient module}
\format{
synapseclient module
}
\usage{
.syn
}
\description{
Main synapseclient module
}
\examples{
\dontrun{
library(synapserprototype)

}
}
\keyword{datasets}
