% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get}
\alias{get}
\title{Gets a Synapse entity}
\usage{
get(synid)
}
\arguments{
\item{synid}{Synapse Id of an entity}
}
\value{
Entity
}
\description{
Gets a Synapse entity
}
\examples{
library(synapserprototype)
entity <- get('syn123')
}
