% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{syn_config}
\alias{syn_config}
\alias{syn_version}
\title{synapseclient configuration information}
\usage{
syn_config()

syn_version()
}
\value{
List with information on the current configuration of synapseclient
  You can determine whether synapseclient was found using the `available`
  member (other members vary depending on whether `available` is `TRUE`
  or `FALSE`)
}
\description{
synapseclient configuration information
}
\keyword{internal}
