% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{syn_login}
\alias{syn_login}
\title{Logs into Synapse.}
\usage{
syn_login(
  email = NULL,
  password = NULL,
  apiKey = NULL,
  rememberMe = F,
  silent = F,
  forced = F,
  authToken = NULL
)
}
\arguments{
\item{email}{a Synapse username (optional, not required if .synapseConfig available)}

\item{password}{a Synapse password (optional, not required if .synapseConfig available)}

\item{apiKey}{Base64 encoded Synapse API key}

\item{rememberMe}{Whether the authentication information should be cached in your operating system's credential storage.}

\item{silent}{Defaults to False.  Suppresses the "Welcome ...!" message.}

\item{forced}{Defaults to False.  Bypass the credential cache if set.}

\item{authToken}{A bearer authorization token, e.g. a personal access token, can be used in lieu of a password or apiKey}
}
\description{
Logs into Synapse.
}
\examples{
library(synapserprototype)
syn_login()
}
