% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{synapserprototype}
\alias{synapserprototype}
\title{synapseclient for R}
\description{
\href{https://github.com/Sage-Bionetworks/synapsePythonClientg}{synapseclient}
is an open source software library that interacts with Synapse,
a collaborative, open-source research platform that allows teams to share
data, track analyses, and collaborate.
}
\details{
For additional documentation on the synapseclient package see
\href{https://python-docs.synapse.org/build/html/index.html}{https://python-docs.synapse.org/build/html/index.html}
}
